requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.1
PyPDF2==3.0.1
python-docx==0.8.11
Markdown==3.4.1
EbookLib==0.17.1
beautifulsoup4==4.12.2
ollama==0.1.6
itsdangerous==2.0.1
click==8.0.4
colorama==0.4.4
MarkupSafe==2.0.1 